backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.9.0
psycopg2-binary==2.9.9
pgvector==0.3.5
google-generativeai==0.8.3
openai==1.54.3
httpx==0.27.2
beautifulsoup4==4.12.3
PyPDF2==3.0.1
pdfplumber==0.11.4
python-multipart==0.0.12
